requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#

backports.shutil-get-terminal-size==1.0.0  # via ipython
click==6.6                # via pip-tools
decorator==4.0.9          # via ipython, traitlets
entrypoints==0.2.1        # via nbconvert
first==2.0.1              # via pip-tools
ipykernel==4.3.1          # via ipywidgets, jupyter, jupyter-console, notebook, qtconsole
ipython-genutils==0.1.0   # via nbformat, notebook, traitlets
ipython==4.2.0            # via ipykernel, ipywidgets, jupyter-console
ipywidgets==5.1.3         # via jupyter
jinja2==2.8               # via nbconvert, notebook
jsonschema==2.5.1         # via nbformat
jupyter-client==4.2.2     # via ipykernel, jupyter-console, notebook, qtconsole
jupyter-console==4.1.1    # via jupyter
jupyter-core==4.1.0       # via jupyter-client, nbconvert, nbformat, notebook, qtconsole
jupyter==1.0.0
markupsafe==0.23          # via jinja2
mistune==0.7.2            # via nbconvert
nbconvert==4.2.0          # via jupyter, notebook
nbformat==4.0.1           # via nbconvert, notebook
notebook==4.2.0           # via jupyter, widgetsnbextension
pexpect==4.0.1            # via ipython
pickleshare==0.7.2        # via ipython
pillow==3.2.0
pip-tools==1.6.4
ptyprocess==0.5.1         # via pexpect, terminado
pygments==2.1.3           # via nbconvert, qtconsole
pyzmq==15.2.0             # via jupyter-client
qtconsole==4.2.1          # via jupyter
simplegeneric==0.8.1      # via ipython
six==1.10.0               # via pip-tools
terminado==0.6            # via notebook
tornado==4.3              # via ipykernel, notebook, terminado
traitlets==4.2.1          # via ipykernel, ipython, ipywidgets, jupyter-client, jupyter-core, nbconvert, nbformat, notebook, qtconsole
widgetsnbextension==1.2.2  # via ipywidgets

# The following packages are commented out because they are
# considered to be unsafe in a requirements file:
# setuptools                # via ipython
